[FIX] Web: view_form: One2ManyViewManager header
Uses the headless flag at initialization of the widget instead of extending
the ViewManager template (which doesn't handle the header anymore).

Also uses the headless flag to tell the inner view not to put its
buttons (use case: kanban) or pager (use case: list) in the header as it
is hidden.

Note: the behavior for list view is unchanged and has a bug. If there are
more then 80 records to display, only the 80th first are display, and as
there is no pager, there is no way to see the remaining ones. A fix for
this would be to use a light HeaderWidget in One2Many views as well.
With this, there won't be duplicated .oe-'view'-buttons/sidebar/pager
anymore.