[IMP] fields: move field.digits and column.digits to the registry
The computed value of digits is now stored in the registry instead of the field
itself.  As a consequence, float columns can be shared among registries.

Note that field.digits and column.digits are set up separately, in order to
avoid the second one being done twice.

This is a first step to make fields shareable across registries: store
database-specific metadata in the registry itself.