[FIX] fields: fix _column.new() by relying on to_field_args()
The former implementation was based on the ill-defined method same_parameters()
that compares arguments based on a heuristic.  That method is now broken, since
it accesses the property 'digits' on float fields before digits are set up.
Instead, the new implementation creates a column and checks whether it is
equivalent to self.