[WIP] Web: ActionManager and ControlPanel
Introduction of State classes to facilitate the way of restoring
previously rendered states in ControlPanel. The notion of widgets
previously used in the ActionManager is now encapsulated in the notion
of states.

There are still some issues to fix, mainly with the breadcrumbs (server
error, view is not displayed...). It seems that the event thrown when
clicking on the breadcrumb is wrong.