[WIP] Web: views.js: breadcrumbs bug fix
Fixes the bug that the view wasn't displayed when clicking on
breadcrumbs (due to the way we prevented to left stuff in the control
panel history). There is now a flag in the set_state function signature
s.t. iff true, we store the content in the history.

Also fixes a bug with the switch_view events thrown by the CP as they
were intercepted by all VMs. Those events are now listened by the AM
which forwards them to the active inner_widget.