[FIX] account.move.line: only significant changes should trigger move invalidation
Many trivial changes to a move line, such as the
"blocked" flag for litigation cannot affect the
balancing status of the move, and should not cause
the move to be (re)validated.
This means it will be possible to change trivial
fields even on moves on a closed fiscal year,
for example.