[FIX] calendar: loading problem with some locales
When you try to load calendar_demo.xml, the loading of demo event
`calendar_event_1` fails with fr_BE locale (and probably some others).

This is due to the rendering of the mail template
`calendar_template_meeting_invitation` (which is used when you create
an event).

This template used a method `get_interval()` in calendar.py, which
does not always return Unicode strings. As jinja2 always expects
Unicode strings, the rendering failed.

The bug is fixed by decoding cleanly the bytestrings returned by the
`locale` Python lib.

PS: Yeah, we love Python2's management of encodings and Unicode...